requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
passlib
bcrypt
pydantic[email]
PyJWT
